chore: separar os estilos do arquivo
